requirements.txt
Django
Fabric
Jinja2
MarkupSafe
Pillow
Pygments
Sphinx
Whoosh
amqp
anyjson
billiard
celery
django-appconf
django-autoslug
django-celery
django-compressor
django-cors-headers
django-guardian
django-gunicorn
django-haystack
django-kombu
django-libsass
django-mail-queue
django-mptt
django-polymorphic
djangorestframework
djangorestframework-jwt
django-registration-rest-framework
django-celery-email
django-filter
docutils
ecdsa
factory-boy
gunicorn
kombu
libsass
paramiko
pycrypto
pyparsing
python-magic
python-ptrace
pytz
requests
six
sorl-thumbnail
django-debug-toolbar
django-statsd-mozilla
django-social-auth
